Include Permutation class into header file
